+++ excel/src/test/java/org/apache/metamodel/excel/ExcelDataContextTest.java
@@ -18,9 +18,12 @@
  */
 package org.apache.metamodel.excel;
 
+import static org.junit.Assert.*;
+
 import java.io.File;
 import java.util.Arrays;
 import java.util.List;
+import java.util.stream.IntStream;
 
 import org.apache.metamodel.DataContext;
 import org.apache.metamodel.MetaModelHelper;
@@ -30,20 +33,30 @@ import org.apache.metamodel.data.DataSet;
 import org.apache.metamodel.data.Row;
 import org.apache.metamodel.data.Style;
 import org.apache.metamodel.data.StyleBuilder;
+import org.apache.metamodel.insert.InsertInto;
 import org.apache.metamodel.query.Query;
 import org.apache.metamodel.schema.Column;
+import org.apache.metamodel.schema.ColumnType;
 import org.apache.metamodel.schema.Schema;
 import org.apache.metamodel.schema.Table;
 import org.apache.metamodel.schema.naming.CustomColumnNamingStrategy;
+import org.apache.metamodel.update.Update;
 import org.apache.metamodel.util.DateUtils;
 import org.apache.metamodel.util.FileHelper;
 import org.apache.metamodel.util.FileResource;
 import org.apache.metamodel.util.Month;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.TemporaryFolder;
+import org.junit.rules.TestName;
 
-import junit.framework.TestCase;
+public class ExcelDataContextTest {
+    @Rule
+    public TemporaryFolder folder = new TemporaryFolder();
 
-public class ExcelDataContextTest extends TestCase {
-
+    @Rule
+    public TestName testName = new TestName();
+    
     /**
      * Creates a copy of a particular file - to avoid changing of Excel files
      * under source control
@@ -53,12 +66,13 @@ public class ExcelDataContextTest extends TestCase {
      */
     private File copyOf(String path) {
         final File srcFile = new File(path);
-        final File destFile = new File("target/" + getName() + "-" + srcFile.getName());
+        final File destFile = new File(folder.getRoot(), testName.getMethodName() + "-" + srcFile.getName());
         FileHelper.copy(srcFile, destFile);
         return destFile;
     }
 
-    public void testErrornousConstructors() throws Exception {
+    @Test
+    public void testErroneousConstructors() throws Exception {
         try {
             new ExcelDataContext(null);
             fail("Exception expected");
@@ -75,6 +89,7 @@ public class ExcelDataContextTest extends TestCase {
         }
     }
 
+    @Test
     public void testEmptyFile() throws Exception {
         File file = copyOf("src/test/resources/empty_file.xls");
         ExcelDataContext dc = new ExcelDataContext(file);
@@ -89,6 +104,7 @@ public class ExcelDataContextTest extends TestCase {
         assertSame(file, ((FileResource) dc.getResource()).getFile());
     }
 
+    @Test
     public void testEmptyFileNoHeaderLine() throws Exception {
         DataContext dc = new ExcelDataContext(copyOf("src/test/resources/empty_file.xls"), new ExcelConfiguration(
                 ExcelConfiguration.NO_COLUMN_NAME_LINE, false, false));
@@ -99,6 +115,7 @@ public class ExcelDataContextTest extends TestCase {
         assertEquals(0, table.getColumnCount());
     }
 
+    @Test
     public void testUnexistingHeaderLine() throws Exception {
         DataContext dc = new ExcelDataContext(copyOf("src/test/resources/xls_people.xls"), new ExcelConfiguration(20,
                 true, false));
@@ -109,6 +126,7 @@ public class ExcelDataContextTest extends TestCase {
         assertEquals(0, table.getColumnCount());
     }
 
+    @Test
     public void testSkipEmptyColumns() throws Exception {
         ExcelConfiguration conf = new ExcelConfiguration(ExcelConfiguration.DEFAULT_COLUMN_NAME_LINE, true, true);
         ExcelDataContext dc = new ExcelDataContext(copyOf("src/test/resources/skipped_lines.xlsx"), conf);
@@ -120,6 +138,7 @@ public class ExcelDataContextTest extends TestCase {
         assertEquals("1", ds.getRow().getValue(0));
     }
 
+    @Test
     public void testDontSkipEmptyLinesNoHeader() throws Exception {
         ExcelConfiguration conf = new ExcelConfiguration(ExcelConfiguration.NO_COLUMN_NAME_LINE, false, true);
         ExcelDataContext dc = new ExcelDataContext(copyOf("src/test/resources/skipped_lines.xlsx"), conf);
@@ -144,6 +163,7 @@ public class ExcelDataContextTest extends TestCase {
         assertEquals("1", ds.getRow().getValue(0));
     }
 
+    @Test
     public void testDontSkipEmptyLinesAbsoluteHeader() throws Exception {
         ExcelConfiguration conf = new ExcelConfiguration(6, false, true);
         ExcelDataContext dc = new ExcelDataContext(copyOf("src/test/resources/skipped_lines.xlsx"), conf);
@@ -157,6 +177,7 @@ public class ExcelDataContextTest extends TestCase {
         assertEquals("1", ds.getRow().getValue(0));
     }
 
+    @Test
     public void testInvalidFormula() throws Exception {
         ExcelDataContext dc = new ExcelDataContext(copyOf("src/test/resources/invalid_formula.xls"));
         Table table = dc.getDefaultSchema().getTables().get(0);
@@ -194,6 +215,7 @@ public class ExcelDataContextTest extends TestCase {
         ds.close();
     }
 
+    @Test
     public void testEvaluateFormula() throws Exception {
         ExcelDataContext dc = new ExcelDataContext(copyOf("src/test/resources/xls_formulas.xls"));
 
@@ -270,6 +292,7 @@ public class ExcelDataContextTest extends TestCase {
         assertFalse(ds.next());
     }
 
+    @Test
     public void testSingleCellSheet() throws Exception {
         ExcelDataContext dc = new ExcelDataContext(copyOf("src/test/resources/xls_single_cell_sheet.xls"));
 
@@ -284,6 +307,7 @@ public class ExcelDataContextTest extends TestCase {
         assertFalse(ds.next());
     }
 
+    @Test
     public void testOpenXlsxFormat() throws Exception {
         ExcelDataContext dc = new ExcelDataContext(copyOf("src/test/resources/Spreadsheet2007.xlsx"));
         Schema schema = dc.getDefaultSchema();
@@ -308,6 +332,7 @@ public class ExcelDataContextTest extends TestCase {
         assertEquals("[bar, 4, 2010-01-04 00:00:00]", Arrays.toString(objectArrays.get(3)));
     }
 
+    @Test
     public void testConfigurationWithoutHeader() throws Exception {
         File file = copyOf("src/test/resources/xls_people.xls");
         DataContext dc = new ExcelDataContext(file, new ExcelConfiguration(ExcelConfiguration.NO_COLUMN_NAME_LINE,
@@ -331,6 +356,7 @@ public class ExcelDataContextTest extends TestCase {
         assertFalse(dataSet.next());
     }
 
+    @Test
     public void testConfigurationNonDefaultColumnNameLineNumber() throws Exception {
         File file = copyOf("src/test/resources/xls_people.xls");
         DataContext dc = new ExcelDataContext(file, new ExcelConfiguration(2, true, true));
@@ -352,6 +378,7 @@ public class ExcelDataContextTest extends TestCase {
         assertFalse(dataSet.next());
     }
 
+    @Test
     public void testGetSchemas() throws Exception {
         File file = copyOf("src/test/resources/xls_people.xls");
         DataContext dc = new ExcelDataContext(file);
@@ -373,6 +400,7 @@ public class ExcelDataContextTest extends TestCase {
         assertEquals("age", columns[3].getName());
     }
 
+    @Test
     public void testMaterializeTable() throws Exception {
         File file = copyOf("src/test/resources/xls_people.xls");
         ExcelDataContext dc = new ExcelDataContext(file);
@@ -395,6 +423,7 @@ public class ExcelDataContextTest extends TestCase {
         assertNull(dataSet.getRow());
     }
 
+    @Test
     public void testMissingValues() throws Exception {
         File file = copyOf("src/test/resources/xls_missing_values.xls");
         DataContext dc = new ExcelDataContext(file);
@@ -419,6 +448,7 @@ public class ExcelDataContextTest extends TestCase {
         assertFalse(ds.next());
     }
 
+    @Test
     public void testMissingColumnHeader() throws Exception {
         File file = copyOf("src/test/resources/xls_missing_column_header.xls");
         DataContext dc = new ExcelDataContext(file);
@@ -443,6 +473,7 @@ public class ExcelDataContextTest extends TestCase {
         assertFalse(ds.next());
     }
 
+    @Test
     public void testXlsxFormulas() throws Exception {
         File file = copyOf("src/test/resources/formulas.xlsx");
         ExcelDataContext dc = new ExcelDataContext(file);
@@ -519,6 +550,7 @@ public class ExcelDataContextTest extends TestCase {
         assertFalse(ds.next());
     }
 
+    @Test
     public void testTicket99defect() throws Exception {
         File file = copyOf("src/test/resources/ticket_199_inventory.xls");
         DataContext dc = new ExcelDataContext(file);
@@ -540,6 +572,7 @@ public class ExcelDataContextTest extends TestCase {
                 Arrays.toString(table.getColumns().toArray()));
     }
 
+    @Test
     public void testInsertInto() throws Exception {
         final File file = copyOf("src/test/resources/xls_people.xls");
 
@@ -594,12 +627,14 @@ public class ExcelDataContextTest extends TestCase {
         assertFalse(ds.next());
     }
 
+    @Test
     public void testCreateTableXls() throws Exception {
         // run the same test with both XLS and XLSX (because of different
         // workbook implementations)
         runCreateTableTest(new File("target/xls_people_created.xls"));
     }
 
+    @Test
     public void testCreateTableXlsx() throws Exception {
         // run the same test with both XLS and XLSX (because of different
         // workbook implementations)
@@ -697,6 +732,7 @@ public class ExcelDataContextTest extends TestCase {
         assertEquals(2, dc.getDefaultSchema().getTableCount());
     }
 
+    @Test
     public void testGetStyles() throws Exception {
         DataContext dc = new ExcelDataContext(copyOf("src/test/resources/styles.xlsx"));
         Table table = dc.getDefaultSchema().getTables().get(0);
@@ -760,6 +796,7 @@ public class ExcelDataContextTest extends TestCase {
      * Tests that you can execute a query on a ExcelDataContext even though the
      * schema has not yet been (explicitly) loaded.
      */
+    @Test
     public void testExecuteQueryBeforeLoadingSchema() throws Exception {
         // first use one DataContext to retreive the schema/table/column objects
         ExcelDataContext dc1 = new ExcelDataContext(copyOf("src/test/resources/Spreadsheet2007.xlsx"));
@@ -777,6 +814,7 @@ public class ExcelDataContextTest extends TestCase {
         ds.close();
     }
 
+    @Test
     public void testCustomColumnNames() throws Exception {
         final String firstColumnName = "first";
         final String secondColumnName = "second";
@@ -791,5 +829,170 @@ public class ExcelDataContextTest extends TestCase {
         assertNotNull(table.getColumnByName(firstColumnName));
         assertNotNull(table.getColumnByName(secondColumnName));
         assertNotNull(table.getColumnByName(thirdColumnName));
+    }
+
+    @Test
+    public void testDetectingDifferentDataTypesInXls() throws Exception {
+        detectingDataTypes("src/test/resources/different_datatypes.xls");
+    }
+
+    @Test
+    public void testDifferentDataTypesInXlsx() throws Exception {
+        detectingDataTypes("src/test/resources/different_datatypes.xlsx");
+    }
+
+    private void detectingDataTypes(final String file) {
+        final DataContext dataContext = new ExcelDataContext(copyOf(file), new ExcelConfiguration(
+                ExcelConfiguration.DEFAULT_COLUMN_NAME_LINE, null, true, false, true, 19));
+
+        final Schema schema = dataContext.getDefaultSchema();
+        assertEquals(2, schema.getTableCount());
+
+        final Table table = schema.getTables().get(0);
+        assertEquals("INTEGER", table.getColumns().get(0).getName());
+        assertEquals(ColumnType.INTEGER, table.getColumns().get(0).getType());
+        assertEquals("TEXT", table.getColumns().get(1).getName());
+        assertEquals(ColumnType.STRING, table.getColumns().get(1).getType());
+        assertEquals("FORMULA", table.getColumns().get(2).getName());
+        assertEquals(ColumnType.INTEGER, table.getColumns().get(2).getType());
+        assertEquals("MIXING_DOUBLE_AND_INT", table.getColumns().get(3).getName());
+        assertEquals(ColumnType.DOUBLE, table.getColumns().get(3).getType());
+        assertEquals("MIXING_OTHER_DATATYPES", table.getColumns().get(4).getName());
+        assertEquals(ColumnType.STRING, table.getColumns().get(4).getType());
+        final DataSet countDataSet = dataContext.query().from(table).selectCount().execute();
+        assertTrue(countDataSet.next());
+        assertEquals(20L, countDataSet.getRow().getValue(0));
+        assertFalse(countDataSet.next());
+    }
+
+    @Test
+    public void testCellValueWithWrongDatatypeIsSetToNull() {
+        // Unless Integers and Doubles are mixed all other incorrect values will be converted to null with a warning
+        final DataContext dataContext = new ExcelDataContext(copyOf("src/test/resources/different_datatypes.xls"),
+                new ExcelConfiguration(ExcelConfiguration.DEFAULT_COLUMN_NAME_LINE, null, true, false, true, 19));
+        final Table table = dataContext.getDefaultSchema().getTables().get(0);
+        final DataSet testWrongDatatypeDataSet = dataContext
+                .query()
+                .from(table)
+                .select("MIXING_DOUBLE_AND_INT")
+                .execute();
+        IntStream.range(0, 19).forEach(i -> {
+            assertTrue(testWrongDatatypeDataSet.next());
+            assertNotNull(testWrongDatatypeDataSet.getRow().getValue(0));
+        });
+        assertTrue(testWrongDatatypeDataSet.next());
+        assertNull(testWrongDatatypeDataSet.getRow().getValue(0));
+        assertFalse(testWrongDatatypeDataSet.next());
+    }
+
+    @Test
+    public void testDetectingDataTypeNotSkippingLinesAndColumnsUsingNameLine() {
+        final ExcelDataContext dataContext = new ExcelDataContext(copyOf("src/test/resources/skipped_lines.xlsx"),
+                new ExcelConfiguration(6, null, false, false, true, 3));
+        final Table table = dataContext.getDefaultSchema().getTables().get(0);
+        assertEquals(ColumnType.STRING, table.getColumns().get(0).getType());
+        assertEquals(ColumnType.INTEGER, table.getColumns().get(6).getType());
+        assertEquals(ColumnType.INTEGER, table.getColumns().get(7).getType());
+    }
+
+    @Test
+    public void testDetectingDataTypeNotSkippingLinesAndColumnsNoNameLine() {
+        final ExcelDataContext dataContext = new ExcelDataContext(copyOf("src/test/resources/skipped_lines.xlsx"),
+                new ExcelConfiguration(ExcelConfiguration.NO_COLUMN_NAME_LINE, null, false, false, true, 3));
+        final Table table = dataContext.getDefaultSchema().getTables().get(0);
+        assertEquals(ColumnType.STRING, table.getColumns().get(0).getType());
+        assertEquals(ColumnType.INTEGER, table.getColumns().get(6).getType());
+        assertEquals(ColumnType.INTEGER, table.getColumns().get(7).getType());
+    }
+
+    @Test
+    public void testDetectingDataTypeSkippingLinesAndColumnsUsingNameLine() {
+        final ExcelDataContext dataContext = new ExcelDataContext(copyOf("src/test/resources/skipped_lines.xlsx"),
+                new ExcelConfiguration(ExcelConfiguration.DEFAULT_COLUMN_NAME_LINE, null, true, true, true, 3));
+        final Table table = dataContext.getDefaultSchema().getTables().get(0);
+        assertEquals(ColumnType.INTEGER, table.getColumns().get(0).getType());
+        assertEquals(ColumnType.INTEGER, table.getColumns().get(1).getType());
+    }
+
+    @Test
+    public void testDetectingDataTypeSkippingLinesAndColumnsNoNameLine() {
+        final ExcelDataContext dataContext = new ExcelDataContext(copyOf("src/test/resources/skipped_lines.xlsx"),
+                new ExcelConfiguration(ExcelConfiguration.NO_COLUMN_NAME_LINE, null, true, true, true, 3));
+        final Table table = dataContext.getDefaultSchema().getTables().get(0);
+        assertEquals(ColumnType.INTEGER, table.getColumns().get(0).getType());
+        assertEquals(ColumnType.INTEGER, table.getColumns().get(1).getType());
+    }
+
+    @Test
+    public void testToMuchNumberOfLinesToScan() throws Exception {
+        final ExcelDataContext dataContext = new ExcelDataContext(copyOf("src/test/resources/skipped_lines.xlsx"),
+                new ExcelConfiguration(ExcelConfiguration.DEFAULT_COLUMN_NAME_LINE, null, true, true, true, 4));
+
+        final Table table = dataContext.getDefaultSchema().getTables().get(0);
+        assertEquals("hello", table.getColumns().get(0).getName());
+        assertEquals("world", table.getColumns().get(1).getName());
+        assertEquals(ColumnType.INTEGER, table.getColumns().get(0).getType());
+        assertEquals(ColumnType.INTEGER, table.getColumns().get(1).getType());
+        final DataSet dataSet = dataContext.query().from(table).selectCount().execute();
+        assertTrue(dataSet.next());
+        assertEquals(3L, dataSet.getRow().getValue(0));
+    }
+
+    @Test
+    public void testToMissingValueDoesntEffectDetectedType() throws Exception {
+        final ExcelDataContext dataContext = new ExcelDataContext(copyOf("src/test/resources/xls_missing_values.xls"),
+                new ExcelConfiguration(ExcelConfiguration.DEFAULT_COLUMN_NAME_LINE, null, true, false, true, 3));
+
+        final Table table = dataContext.getDefaultSchema().getTables().get(0);
+        final Column columnB = table.getColumns().get(1);
+        assertEquals("b", columnB.getName());
+        assertEquals(ColumnType.INTEGER, columnB.getType());
+        final DataSet dataSetColumnB = dataContext.query().from(table).select(columnB).execute();
+        assertTrue(dataSetColumnB.next());
+        assertTrue(dataSetColumnB.next());
+        assertNull(dataSetColumnB.getRow().getValue(0));
+
+        final Column columnD = table.getColumns().get(3);
+        assertEquals("d", columnD.getName());
+        assertEquals(ColumnType.INTEGER, columnD.getType());
+        final DataSet dataSetColumnD = dataContext.query().from(table).select(columnD).where(columnD).eq(12).execute();
+        assertTrue(dataSetColumnD.next());
+        assertEquals(12, dataSetColumnD.getRow().getValue(0));
+    }
+
+    @Test
+    public void testInsertingValueOfValidColumnType() {
+        final ExcelDataContext dataContext = new ExcelDataContext(copyOf("src/test/resources/different_datatypes.xls"),
+                new ExcelConfiguration(ExcelConfiguration.DEFAULT_COLUMN_NAME_LINE, null, true, false, true, 19));
+        final Table table = dataContext.getDefaultSchema().getTable(0);
+        dataContext.executeUpdate(new InsertInto(table).value("INTEGER", Integer.valueOf(123)));
+        final DataSet dataSet = dataContext.query().from(table).selectAll().where("INTEGER").eq(123).execute();
+        assertTrue(dataSet.next());
+    }
+
+    @Test
+    public void testInsertingValueOfInvalidColumnType() {
+        final ExcelDataContext dataContext = new ExcelDataContext(copyOf("src/test/resources/different_datatypes.xls"),
+                new ExcelConfiguration(ExcelConfiguration.DEFAULT_COLUMN_NAME_LINE, null, true, false, true, 19));
+        final Table table = dataContext.getDefaultSchema().getTable(0);
+        dataContext.executeUpdate(new InsertInto(table).value("INTEGER", "this is not an integer"));
+    }
+
+    @Test
+    public void testUpdatingValueOfValidColumnType() {
+        final ExcelDataContext dataContext = new ExcelDataContext(copyOf("src/test/resources/different_datatypes.xls"),
+                new ExcelConfiguration(ExcelConfiguration.DEFAULT_COLUMN_NAME_LINE, null, true, false, true, 19));
+        final Table table = dataContext.getDefaultSchema().getTable(0);
+        dataContext.executeUpdate(new Update(table).value("INTEGER", 1).value("INTEGER", Integer.valueOf(123)));
+        final DataSet dataSet = dataContext.query().from(table).selectAll().where("INTEGER").eq(123).execute();
+        assertTrue(dataSet.next());
+    }
+
+    @Test
+    public void testUpdatingValueOfInvalidColumnType() {
+        final ExcelDataContext dataContext = new ExcelDataContext(copyOf("src/test/resources/different_datatypes.xls"),
+                new ExcelConfiguration(ExcelConfiguration.DEFAULT_COLUMN_NAME_LINE, null, true, false, true, 19));
+        final Table table = dataContext.getDefaultSchema().getTable(0);
+        dataContext.executeUpdate(new Update(table).value("INTEGER", 1).value("INTEGER", "this is not an integer"));
     }
 }